+++ src/lib/components/modal/Modal.svelte
@@ -12,6 +12,7 @@
     export type ModalProps = {
       open?: boolean;
       escapeKeyClose?: boolean;
+      clickOutsideClose?: boolean;
       disableBodyScroll?: boolean;
       ariaLabel?: string;
       inertId?: string;
@@ -24,7 +25,7 @@
 </script>
 
 <script lang="ts">
-  let {open=$bindable(false), escapeKeyClose=true, disableBodyScroll=true, transitionDuration=config.transitionDuration, inertId, ariaLabel="Modal", overlayClasses="", children, ...props}: ModalProps = $props();
+  let {open=$bindable(false), escapeKeyClose=true, clickOutsideClose=true, disableBodyScroll=true, transitionDuration=config.transitionDuration, inertId, ariaLabel="Modal", overlayClasses="", children, ...props}: ModalProps = $props();
 
   let modalElement: HTMLDivElement | undefined = $state();
 
@@ -50,11 +51,17 @@
     }
   };
 
+  function handleOverlayClick() {
+    if(clickOutsideClose) {
+      closeModal();
+    }
+  }
+
 </script>
 
 {#if open}
 <Portal class="fixed z-50">
-  <Overlay {transitionDuration} {disableBodyScroll} class={overlayClasses} onclick={closeModal} />
+  <Overlay {transitionDuration} {disableBodyScroll} class={overlayClasses} onclick={handleOverlayClick} />
   <div bind:this={modalElement} use:trapFocus role="dialog" aria-modal="true" aria-label={ariaLabel} tabindex={open ? 0 : -1} aria-hidden={!open}
     class={twMerge("fixed w-full max-w-2xl top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 h-6/12 focus:outline-none bg-white rounded-primary", props.class)}
     use:onKeydown={{key: "Escape", callback: handleKeydown}}